+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - Improved the `show` command to make it easier to check if packages are properly installed.
 - The `script` command has been deprecated, use `run` instead.
+- Improved support for private repositories.
 - Expanded version constraints now keep the original version's precision.
 
 ### Fixed